chore: apply pull request suggestions
