+++ events/gateway_ratelimited_events.go
@@ -2,8 +2,8 @@ package events
 
 import "github.com/disgoorg/disgo/gateway"
 
-// RateLimited indicates we received a RateLimited from the gateway.Gateway
-type RateLimited struct {
+// GatewayRateLimited indicates we received a RateLimited from the gateway.Gateway
+type GatewayRateLimited struct {
 	*GenericEvent
 	gateway.EventRateLimited
 }
